password

+++ app/Http/Controllers/PemindahBukuanController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Support\Arr;
 use Illuminate\Http\Request;
 use App\Models\PemindahBukuan;
 use Maatwebsite\Excel\Facades\Excel;
@@ -34,6 +33,11 @@ class PemindahBukuanController extends Controller
 
     public function upload(Request $request)
     {
+        if ($request->password != 'password') {
+            return redirect()
+                ->route('welcome');
+        }
+
         PemindahBukuan::truncate();
 
         Excel::import(new PemindahBukuansImport, $request->bps);
@@ -46,6 +50,17 @@ class PemindahBukuanController extends Controller
     public function detail(PemindahBukuan $pemindahBukuan)
     {
         return view('detail', compact('pemindahBukuan'));
+    }
+
+    public function passwordAll(Request $request)
+    {
+        if ($request->password != 'password') {
+            return redirect()
+                ->route('welcome');
+        }
+
+        return redirect()
+            ->route('all');
     }
 
     public function all()
